requirements-gpu.txt
tensorflow-gpu>=2.12.0
opencv-python>=4.2.0.32
lxml
tqdm

-e .
